requirements.txt
beautifulsoup4>=4.12.2
lxml>=4.9.3
playwright>=1.47.0
python-telegram-bot>=20.0
python-dotenv>=1.0.0
